:sparkles: Leaderboard commands

+++ src/commands/administration/resetuser.ts
@@ -16,18 +16,33 @@ export default class ResetUser extends AdminCommand {
     userData: DocumentType<DbUser>,
     guildData: DocumentType<DbGuild>
   ) {
-    const user = message.mentions.users.first();
+    const userId = message.mentions.users.first()?.id || args[0];
+    if (!userId)
+      return message.channel.send(
+        embeds.error(
+          `Please send the id or tag the user you would like to reset!`
+        )
+      );
+    const user = this.client.users.resolve(userId);
     if (!user)
       return message.channel.send(
-        embeds.error(`Please tag the user you would like to reset!`)
+        embeds.error(`The user id or tag you provided is invalid.`)
       );
 
-    await UserModel.deleteOne({
+    const deletedData = await UserModel.deleteOne({
       userId: user.id,
     });
 
-    message.channel.send(
-      embeds.normal(`User Profile Reset`, `${user} has been completely reset.`)
-    );
+    if (deletedData.deletedCount)
+      return message.channel.send(
+        embeds.normal(
+          `User Profile Reset`,
+          `${user} has been completely reset.`
+        )
+      );
+    else
+      message.channel.send(
+        embeds.error(`The user does not have a registered profile.`)
+      );
   }
 }
